Rename liferay mock to oauth2_provider mock
