+++ src/layman/common/prime_db_schema/publications_test.py
@@ -11,7 +11,7 @@ from . import publications, workspaces, users
 DB_SCHEMA = settings.LAYMAN_PRIME_SCHEMA
 
 userinfo_baseline = {"issuer_id": 'mock_test_publications_test',
-                     "claims": {"email": "test@liferay.com",
+                     "claims": {"email": "test@oauth2.org",
                                 "preferred_username": 'test_preferred',
                                 "name": "test ensure user",
                                 "given_name": "test",
